Primitive E-Tag support.
Allow for use of fixed string E-Tags or E-Tags generated from the data sent through the stream. The API for building the E-Tag exactly matches node's `crypto` hash API so that it is easy to use functions from that family.

The new options object looks something like:

```javascript
{
	ETag: crypto.createHash('sha1')
}
```

The default value is to use an MD5 hash, which matches Amazon's default behavior in the general case (though they use something different for multipart uploads). Perhaps in the future support will be added to match their esoteric E-Tag default, but this behavior is more consistent and predictable so it's what we're sticking with. Amazon's choice may have been influenced by the desire to resume uploads later and not have to recalculate the hash from the beginning of the file; not only does that use case not apply here (yet?) but recalcuating the hash from the start of the file is desirable anyway since it ensures file integrity.