Fix deadlock happens if a big file is uploaded
in the write stream, the stream uncork happens before the actual write method ends, and so the data doesn't get written completely. When this happen, the drain event doesn't get fired and so who ever writes to the stream waits indefnitely for the drain event.
The fix was to move the uncork, cork calls to the callback of the stream.write method